move stub files under governance directory
