+++ superset-frontend/pinterest-plugins/src/governance/pinterestTieringInfoModal.stub.tsx
